+++ VoiceInk/Views/PermissionsView.swift
@@ -2,7 +2,6 @@ import SwiftUI
 import AVFoundation
 import Cocoa
 import CoreGraphics
-import PermissionFlow
 
 @MainActor
 class PermissionManager: ObservableObject {
@@ -13,7 +12,6 @@ class PermissionManager: ObservableObject {
     @Published var isKeyboardShortcutSet = false
     @Published var inputMonitoringNeedsRelaunch = false
     @Published var screenRecordingNeedsRelaunch = false
-    private let permissionFlowGuide = PermissionFlowGuide()
     private var permissionRefreshTimer: Timer?
     private var permissionRefreshPollsRemaining = 0
     
@@ -78,7 +76,7 @@ class PermissionManager: ObservableObject {
     
     func requestScreenRecordingPermission() {
         screenRecordingNeedsRelaunch = false
-        permissionFlowGuide.open(.screenRecording)
+        PermissionGrantCoordinator.grantScreenRecording()
         startPermissionRefreshPolling()
         markRelaunchNeededIfPermissionStillInactive(.screenRecording)
     }
@@ -107,33 +105,17 @@ class PermissionManager: ObservableObject {
 
     func requestInputMonitoringPermission() {
         inputMonitoringNeedsRelaunch = false
-        let granted = ShortcutMonitor.requestListenEventAccess()
-        isInputMonitoringEnabled = granted || ShortcutMonitor.preflightListenEventAccess()
-        permissionFlowGuide.open(.inputMonitoring)
+        PermissionGrantCoordinator.grantInputMonitoring { [weak self] granted in
+            self?.isInputMonitoringEnabled = granted
+        }
         startPermissionRefreshPolling()
         markRelaunchNeededIfPermissionStillInactive(.inputMonitoring)
     }
 
-    nonisolated static func openInputMonitoringSettings() {
-        Task { @MainActor in
-            if let url = URL(string: "x-apple.systempreferences:com.apple.preference.security?Privacy_ListenEvent") {
-                NSWorkspace.shared.open(url)
-            }
-        }
-    }
-
-    nonisolated static func openAccessibilitySettings() {
-        Task { @MainActor in
-            if let url = URL(string: "x-apple.systempreferences:com.apple.preference.security?Privacy_Accessibility") {
-                NSWorkspace.shared.open(url)
-            }
-        }
-    }
-
     func requestAccessibilityPermission() {
-        let options: NSDictionary = [kAXTrustedCheckOptionPrompt.takeUnretainedValue() as String: true]
-        _ = AXIsProcessTrustedWithOptions(options)
-        permissionFlowGuide.open(.accessibility)
+        PermissionGrantCoordinator.grantAccessibility { [weak self] granted in
+            self?.isAccessibilityEnabled = granted
+        }
         startPermissionRefreshPolling()
     }